BO/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1